requirements.txt
google-genai

scikit-learn

numpy
pandas


matplotlib
seaborn
plotly


nbformat

python-dotenv

fastapi
uvicorn
python-multipart 
jinja2
requests